@@ -13,46 +13,44 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^4.4.1",
-    "@angular/cdk": "^2.0.0-beta.10",
-    "@angular/common": "^4.4.1",
-    "@angular/compiler": "^4.4.1",
-    "@angular/core": "^4.4.1",
-    "@angular/forms": "^4.4.1",
-    "@angular/http": "^4.4.1",
-    "@angular/material": "^2.0.0-beta.10",
-    "@angular/platform-browser": "^4.4.1",
-    "@angular/platform-browser-dynamic": "^4.4.1",
-    "@angular/router": "^4.4.1",
-    "@ngrx/core": "^1.2.0",
-    "@ngrx/store": "^4.0.3",
-    "core-js": "^2.5.1",
-    "rxjs": "^5.4.3",
-    "zone.js": "^0.8.17"
+    "@angular/animations": "^6.0.3",
+    "@angular/cdk": "^6.1.0",
+    "@angular/common": "^6.0.3",
+    "@angular/compiler": "^6.0.3",
+    "@angular/core": "^6.0.3",
+    "@angular/forms": "^6.0.3",
+    "@angular/http": "^6.0.3",
+    "@angular/material": "^6.1.0",
+    "@angular/platform-browser": "^6.0.3",
+    "@angular/platform-browser-dynamic": "^6.0.3",
+    "@angular/router": "^6.0.3",
+    "core-js": "^2.5.4",
+    "rxjs": "^6.0.0",
+    "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "^1.4.2",
-    "@angular/compiler-cli": "^4.4.1",
-    "@angular/language-service": "^4.4.1",
-    "@types/jasmine": "^2.6.0",
-    "@types/jasminewd2": "^2.0.3",
-    "@types/node": "^8.0.28",
-    "codelyzer": "^3.2.0",
+    "@angular-devkit/build-angular": "~0.6.5",
+    "@angular/cli": "~6.0.5",
+    "@angular/compiler-cli": "^6.0.3",
+    "@angular/language-service": "^6.0.3",
+    "@types/jasmine": "~2.8.6",
+    "@types/jasminewd2": "~2.0.3",
+    "@types/node": "~8.9.4",
+    "codelyzer": "~4.2.1",    
     "commitizen": "^2.9.6",
     "cz-conventional-changelog": "^2.0.0",
     "husky": "^0.14.3",
-    "jasmine-core": "^2.8.0",
-    "jasmine-spec-reporter": "^4.2.1",
-    "karma": "^1.7.1",
-    "karma-chrome-launcher": "^2.2.0",
-    "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "^1.2.1",
-    "karma-jasmine": "~1.1.0",
+    "jasmine-core": "~2.99.1",
+    "jasmine-spec-reporter": "~4.2.1",
+    "karma": "~1.7.1",
+    "karma-chrome-launcher": "~2.2.0",
+    "karma-coverage-istanbul-reporter": "~1.4.2",
+    "karma-jasmine": "~1.1.1",
     "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.1.2",
-    "ts-node": "^3.3.0",
-    "tslint": "^5.7.0",
-    "typescript": "^2.3.4",
+    "protractor": "~5.3.0",
+    "ts-node": "~5.0.1",
+    "tslint": "~5.9.1",
+    "typescript": "2.7.2",
     "validate-commit-message": "^3.0.1"
   }
 }
