@@ -1,15 +1,12 @@
 import { Injectable } from '@angular/core';
 import { Http, URLSearchParams } from '@angular/http';
-import 'rxjs/add/operator/map';
+import { map } from 'rxjs/operators';
 
 @Injectable()
 export class GiphyService {
+  constructor(private http: Http) {}
 
-  constructor(
-    private http: Http
-  ) {}
-
-  search (term: string) {
+  search(term: string) {
     let search = new URLSearchParams();
     search.set('api_key', 'dc6zaTOxFJmzC');
     search.set('q', term);
@@ -18,14 +15,15 @@ export class GiphyService {
 
     return this.http
       .get('http://api.giphy.com/v1/gifs/search', { search })
-      .map(res => res.json()['data'])
-      .map(image =>
-        image.map(g => ({
-          still: g.images.fixed_width_still,
-          small: g.images.fixed_width_small,
-          large: g.images.downsized_large
-        })))
-      ;
+      .pipe(
+        map(res => res.json()['data']),
+        map(image =>
+          image.map(g => ({
+            still: g.images.fixed_width_still,
+            small: g.images.fixed_width_small,
+            large: g.images.downsized_large
+          }))
+        )
+      );
   }
-
 }
