@@ -6,6 +6,6 @@ export class NgPlaygroundPage {
   }
 
   getParagraphText() {
-    return element(by.css('app-root h1')).getText();
+    return element(by.css('app-root mat-toolbar span')).getText();
   }
 }
